+++ packages/owlbot-bootstrapper/common-container/test/split-repo.test.ts
@@ -39,17 +39,6 @@ describe('SplitRepo class', async () => {
       await execSync(
         `mkdir ${repoToClonePath}; cd ${repoToClonePath}; git init --bare`
       );
-      fs.writeFileSync(
-        `${directoryPath}/${utils.INTER_CONTAINER_VARS_FILE}`,
-        JSON.stringify(
-          {
-            branchName: 'specialName',
-            owlbotYamlPath: 'packages/google-cloud-kms/.OwlBot.yaml',
-          },
-          null,
-          4
-        )
-      );
     } catch (err) {
       if (!(err as any).toString().match(/File exists/)) {
         throw err;
@@ -192,6 +181,17 @@ describe('SplitRepo class', async () => {
   });
 
   it('should create and initialize an empty repo on github, then push to main and create an empty PR', async () => {
+    fs.writeFileSync(
+      `${directoryPath}/${utils.INTER_CONTAINER_VARS_FILE}`,
+      JSON.stringify(
+        {
+          branchName: 'specialName',
+          owlbotYamlPath: 'packages/google-cloud-kms/.OwlBot.yaml',
+        },
+        null,
+        4
+      )
+    );
     const scope = nock('https://api.github.com')
       .post('/orgs/googleapis/repos')
       .reply(201)